complete splitting of media class
