+++ ofscraper/classes/media.py
@@ -1,3 +1,14 @@
+import re
+import aiohttp
+import arrow
+from tenacity import retry,stop_after_attempt,wait_random
+from ..constants import LICENCE_URL
+import ofscraper.utils.args as args_
+import ofscraper.utils.auth as auth
+from mpegdash.parser import MPEGDASHParser
+import ofscraper.constants as constants
+import ofscraper.utils.config as config
+
 class Media():
     def __init__(self, media, count, post):
         self._media = media
