+++ test/db/stories_test.py
@@ -2,7 +2,8 @@ import tempfile
 from ofscraper.db.operations import *
 import pytest
 from test.test_constants import *
-from ofscraper.classes.posts import Post,Media
+from ofscraper.classes.posts import Post
+from ofscraper.classes.media import Media
 
 def test_stories_create(mocker):
     with tempfile.NamedTemporaryFile() as p:
