+++ test/post/archived_post_test.py
@@ -1,5 +1,6 @@
 from test.test_constants import *
-from ofscraper.classes.posts import Post,Media
+from ofscraper.classes.posts import Post
+from ofscraper.classes.media import Media
 import re
 import ofscraper.utils.args as args_
 from pytest_check import check
@@ -234,7 +235,7 @@ def test_archived_text_wordtruncate(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs([])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=["",""]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -260,7 +261,7 @@ def test_archived_text_wordtruncate2(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs([])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=["",""]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -286,7 +287,7 @@ def test_archived_text_wordtruncate3(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs([])
-    mocker.patch('ofscraper.api.posts.config.read_config',return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config',return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=["",""]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -311,7 +312,7 @@ def test_archived_text_wordtruncate4(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs([])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=[]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -337,7 +338,7 @@ def test_archived_text_wordtruncate5(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs([])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=[]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -363,7 +364,7 @@ def test_archived_text_wordtruncate6(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs([])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=[]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -386,7 +387,7 @@ def test_archived_text_lettertruncate(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs(["--letter-count"])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=["",""]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -410,7 +411,7 @@ def test_archived_text_lettertruncate2(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs(["--letter-count"])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=["",""]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -432,7 +433,7 @@ def test_archived_text_lettertruncate3(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs(["--letter-count"])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=["",""]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -457,7 +458,7 @@ def test_archived_text_lettertruncate4(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs(["--letter-count"])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=[]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -480,7 +481,7 @@ def test_archived_text_lettertruncate5(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs(["--letter-count"])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=[]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
@@ -502,7 +503,7 @@ def test_archived_text_lettertruncate6(mocker):
         "filter": FILTER_DEFAULT
     }
     args_.getargs(["--letter-count"])
-    mocker.patch('ofscraper.api.posts.config.read_config', return_value=migrationConfig)
+    mocker.patch('ofscraper.classes.posts.config.read_config', return_value=migrationConfig)
     post=mocker.PropertyMock()
     post.post_media=[]
     mocker.patch.object(Media,"text",new=f"{LONG_STRING}{LONG_STRING}{LONG_STRING}{LONG_STRING}")
